declare dispatch in the components that use it
